Add files via upload
Here is the CSV file of the raw flying etiquette data. I also separately uploaded the pivot table I used to create the chart answering the question, "Is it rude to move to an unsold seat on a plane?" I uploaded both the pivot table and the raw data so that both would be available in the repository.